core: Retrieve the provider config schema during provider eval
This EvalNode is currently used within EvalIf so that we minimize the
impact on the existing codepath when the HCL2 experiment is not in use.
Schema is not required when the main HCL/HIL codepaths are being used,
so we don't want to risk failing here if we aren't going to use the schema
anyway.